re-add old sidebar, nav layouts
