+++ ui/apps/dashboard/src/components/Layout/Layout.tsx
@@ -2,12 +2,9 @@ import { Suspense, type ReactNode } from 'react';
 
 import { type ProfileDisplayType } from '@/queries/server/profile';
 import type { Environment } from '@/utils/environments';
-import BottomBar from './BottomBar';
-import SideBar from './SideBar';
-import TopBar from './TopBar';
+import SideBar from './SideBarV2';
 import { ActiveBanners } from '../ActiveBanners/ActiveBanners';
 import IncidentBanner from '../Incident/IncidentBanner';
-import useOnboardingWidget from '../Onboarding/useOnboardingWidget';
 
 type LayoutProps = {
   collapsed: boolean | undefined;
@@ -22,43 +19,24 @@ export default function Layout({
   profile,
   children,
 }: LayoutProps) {
-  // Hoist the onboarding-widget state so the trigger can live in the org
-  // menu (top bar) while the widget itself renders in the sidebar.
-  const { isWidgetOpen, showWidget, closeWidget } = useOnboardingWidget();
-
   return (
-    <div className="bg-canvasSubtle fixed inset-0 flex flex-col overflow-hidden overscroll-none">
-      <TopBar
-        activeEnv={activeEnv}
-        profile={profile}
-        showOnboardingWidget={showWidget}
-      />
+    <div
+      id="layout-scroll-container"
+      className="fixed z-50 flex h-screen w-full flex-row justify-start overflow-y-scroll overscroll-y-none"
+    >
+      <SideBar activeEnv={activeEnv} collapsed={collapsed} profile={profile} />
 
-      <div className="border-subtle bg-canvasBase shadow-xs mx-3 flex flex-1 flex-row overflow-hidden rounded border-hairline">
-        <SideBar
-          activeEnv={activeEnv}
-          collapsed={collapsed}
-          isWidgetOpen={isWidgetOpen}
-          closeWidget={closeWidget}
-        />
+      <div className="no-scrollbar flex w-full flex-col overflow-x-scroll">
+        <IncidentBanner />
 
-        <div
-          id="layout-scroll-container"
-          className="no-scrollbar flex flex-1 flex-col overflow-auto overscroll-contain"
-        >
-          <IncidentBanner />
+        <ActiveBanners />
 
-          <ActiveBanners />
+        {/* disabled by Dan 11/22/2025 for performance reasons */}
+        <Suspense>{/* <BillingBanner /> */}</Suspense>
+        <Suspense>{/* <ExecutionOverageBanner /> */}</Suspense>
 
-          {/* disabled by Dan 11/22/2025 for performance reasons */}
-          <Suspense>{/* <BillingBanner /> */}</Suspense>
-          <Suspense>{/* <ExecutionOverageBanner /> */}</Suspense>
-
-          {children}
-        </div>
+        {children}
       </div>
-
-      <BottomBar />
     </div>
   );
 }
