+++ ui/apps/dashboard/src/components/Layout/SideBar.tsx
@@ -1,33 +1,37 @@
 import { useEffect, useRef, useState } from 'react';
-import { OptionalTooltip } from '@inngest/components/Tooltip/OptionalTooltip';
-import { RiArrowLeftSLine, RiArrowRightSLine } from '@remixicon/react';
 
+import type { ProfileDisplayType } from '@/queries/server/profile';
 import type { Environment } from '@/utils/environments';
-import NavigationV1 from '../Navigation/Navigation';
-import OnboardingWidgetV1 from '../Navigation/OnboardingWidget';
-import NavigationV2 from '../NavigationV2/Navigation';
-import OnboardingWidgetV2 from '../NavigationV2/OnboardingWidget';
+import Logo from '../Navigation/Logo';
+import Navigation from '../Navigation/Navigation';
+import { Profile } from '../Navigation/Profile';
+import { Integrations } from '../Navigation/Integrations';
+import { Help } from '../Navigation/Help';
+import useOnboardingWidget from '../Onboarding/useOnboardingWidget';
 import SeatOverageWidget from '../SeatOverage/SeatOverageWidget';
-import { useNavigationV2 } from './useNavigationV2';
+import OnboardingWidget from '../Navigation/OnboardingWidget';
+
+// Disable SSR in Onboarding Widget, to prevent hydration errors. It requires windows info
+// const OnboardingWidget = dynamic(() => import('../Navigation/OnboardingWidget'), {
+//   ssr: false,
+// });
+// const SeatOverageWidget = dynamic(() => import('../SeatOverage/SeatOverageWidget'), {
+//   ssr: false,
+// });
 
 export default function SideBar({
   collapsed: serverCollapsed,
   activeEnv,
-  isWidgetOpen,
-  closeWidget,
+  profile,
 }: {
   collapsed: boolean | undefined;
   activeEnv?: Environment;
-  isWidgetOpen: boolean;
-  closeWidget: () => void;
+  profile?: ProfileDisplayType;
 }) {
   const navRef = useRef<HTMLDivElement>(null);
 
-  const navV2 = useNavigationV2();
-  const Navigation = navV2 ? NavigationV2 : NavigationV1;
-  const OnboardingWidget = navV2 ? OnboardingWidgetV2 : OnboardingWidgetV1;
-
   const [collapsed, setCollapsed] = useState<boolean>(serverCollapsed ?? false);
+  const { isWidgetOpen, showWidget, closeWidget } = useOnboardingWidget();
 
   const autoCollapse = () =>
     typeof window !== 'undefined' &&
@@ -48,75 +52,32 @@ export default function SideBar({
     }
   }, [serverCollapsed]);
 
-  const toggleCollapsed = () => {
-    const toggled = !collapsed;
-    setCollapsed(toggled);
-
-    if (typeof window !== 'undefined') {
-      window.cookieStore.set('navCollapsed', toggled ? 'true' : 'false');
-      // some downstream things, like charts, may need to redraw themselves
-      setTimeout(() => window.dispatchEvent(new Event('navToggle')), 200);
-    }
-  };
-
-  // ⌘ B / Ctrl + B toggles the sidebar, except when the user is typing in a
-  // form field — otherwise typing "b" with a modifier in the search bar would
-  // collapse the sidebar from under them.
-  useEffect(() => {
-    function onKeyDown(e: KeyboardEvent) {
-      if (e.key !== 'b' || !(e.metaKey || e.ctrlKey)) return;
-      const el = document.activeElement as HTMLElement | null;
-      if (
-        el &&
-        (el.tagName === 'INPUT' ||
-          el.tagName === 'TEXTAREA' ||
-          el.isContentEditable)
-      ) {
-        return;
-      }
-      e.preventDefault();
-      toggleCollapsed();
-    }
-
-    document.addEventListener('keydown', onKeyDown);
-    return () => document.removeEventListener('keydown', onKeyDown);
-  }, [collapsed]);
-
   return (
     <nav
-      className={`bg-canvasBase border-subtle group relative flex h-full flex-col justify-start py-3 transition-[width] duration-200 ease-out ${
-        collapsed ? 'w-[64px]' : 'w-[200px]'
-      } shrink-0 overflow-visible border-r-hairline`}
+      className={`bg-canvasBase border-subtle group
+         top-0 flex h-screen flex-col justify-start ${
+           collapsed ? 'w-[64px]' : 'w-[224px]'
+         }  sticky z-[51] shrink-0 overflow-visible border-r`}
       ref={navRef}
     >
-      {/* Vertical tab on the divider, hover-revealed. Sits half inside / half
-          outside the sidebar so it reads as part of the right edge itself. */}
-      <OptionalTooltip
-        tooltip={collapsed ? 'Expand sidebar' : 'Collapse sidebar'}
-      >
-        <button
-          type="button"
-          onClick={toggleCollapsed}
-          aria-label={collapsed ? 'Expand sidebar' : 'Collapse sidebar'}
-          className="bg-canvasBase border-subtle text-muted hover:text-basis shadow-xs absolute right-0 top-1/2 z-[70] hidden h-8 w-5 -translate-y-1/2 translate-x-1/2 items-center justify-center rounded-md border-hairline transition-colors group-hover:flex"
-        >
-          {collapsed ? (
-            <RiArrowRightSLine className="h-4 w-4" />
-          ) : (
-            <RiArrowLeftSLine className="h-4 w-4" />
-          )}
-        </button>
-      </OptionalTooltip>
-
+      <Logo
+        collapsed={collapsed}
+        envSlug={activeEnv?.slug ?? 'production'}
+        envName={activeEnv?.name ?? 'Production'}
+        setCollapsed={setCollapsed}
+      />
       <div className="flex grow flex-col justify-between">
         <Navigation collapsed={collapsed} activeEnv={activeEnv} />
 
-        <div className="pl-3 pr-2">
+        <div className="mx-4">
           <SeatOverageWidget collapsed={collapsed} />
           {isWidgetOpen && (
             <OnboardingWidget collapsed={collapsed} closeWidget={closeWidget} />
           )}
+          <Integrations collapsed={collapsed} />
+          <Help collapsed={collapsed} showWidget={showWidget} />
         </div>
+        {profile && <Profile collapsed={collapsed} profile={profile} />}
       </div>
     </nav>
   );
